+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -63,6 +63,7 @@ module Test.Cardano.Ledger.Shelley.ImpTest (
   getByronKeyPair,
   freshSafeHash,
   freshKeyHashVRF,
+  freshBlsKey,
   submitTx,
   submitTx_,
   submitTxAnn,
@@ -112,6 +113,7 @@ module Test.Cardano.Ledger.Shelley.ImpTest (
   expectDelegatedToPool,
   getAccountAddressFor,
   freshUnregisteredAccount,
+  freshStakePool,
   freshPoolParams,
   registerPool,
   registerPoolWithAccountAddress,
@@ -1933,6 +1935,9 @@ freshKeyHashVRF ::
   ImpTestM era (VRFVerKeyHash (r :: KeyRoleVRF))
 freshKeyHashVRF = arbitrary
 
+freshBlsKey :: ImpTestM era BlsKey
+freshBlsKey = arbitrary
+
 -- | Adds a key pair to the keyhash lookup map
 addKeyPair ::
   (HasKeyPairs s, MonadState s m) =>
@@ -2256,6 +2261,16 @@ registerAccountAddress ::
   ) =>
   ImpTestM era AccountAddress
 registerAccountAddress = freshKeyHash >>= registerStakeCredential . KeyHashObj
+
+freshStakePool ::
+  ShelleyEraImp era =>
+  ImpTestM era (StakePoolParams era)
+freshStakePool = do
+  poolId <- freshKeyHash
+  ownerKeyHash <- freshKeyHash
+  let owner = KeyHashObj ownerKeyHash
+  ownerAccountAddress <- registerStakeCredential owner
+  freshPoolParams poolId ownerAccountAddress
 
 freshPoolParams ::
   ShelleyEraImp era =>