+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Arbitrary.hs
@@ -42,6 +42,7 @@ import Cardano.Crypto.DSIGN (
   BLS12381MinSigDSIGN,
   DSIGNAggregatable (PossessionProofDSIGN, createPossessionProofDSIGN),
   DSIGNAlgorithm (genKeyDSIGNWithContext),
+  deriveVerKeyDSIGN,
   seedSizeDSIGN,
  )
 import Cardano.Crypto.DSIGN.BLS12381.Internal (minSigPoPDST)
@@ -503,7 +504,15 @@ instance Arbitrary PoolMetadata where
   arbitrary = PoolMetadata <$> arbitrary <*> arbitrary
 
 instance Arbitrary BlsKey where
-  arbitrary = BlsKey <$> arbitrary <*> arbitrary
+  arbitrary = do
+    signingKey <- arbitrary
+    pure $
+      BlsKey
+        (deriveVerKeyDSIGN signingKey)
+        ( createPossessionProofDSIGN
+            minSigPoPDST
+            signingKey
+        )
 
 instance Arbitrary BlsKeyState where
   arbitrary = BlsKeyState <$> arbitrary <*> arbitrary